pyproject.toml
[tool.poetry]
name = "tweetsulike"
version = "0.1.0"
description = ""
authors = ["Yuki Nagae <yuki.nagae1130@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
tweepy = "^3.7"
pip = "^19.1"
python-dotenv = "^0.10.3"
Flask = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
